+++ src/engine/src/rabbitescape/engine/behaviours/Behaviour.java
@@ -9,14 +9,13 @@ import rabbitescape.engine.*;
 import rabbitescape.engine.ChangeDescription.State;
 
 
-public abstract class Behaviour implements RabbitNotifier
+public abstract class Behaviour extends ObservableSubject
 {
     /**
-     * Subclasses examine the rabbit's situation using BehaviourTools and
-     * return the state (see ChangeDescription) for the next time step.
-     * This method may return null indicating that a different Behaviour
-     * must take over.
-     * Note that the state determines the animation used.
+     * Subclasses examine the rabbit's situation using BehaviourTools and return
+     * the state (see ChangeDescription) for the next time step. This method may
+     * return null indicating that a different Behaviour must take over. Note
+     * that the state determines the animation used.
      */
     public abstract State newState( BehaviourTools t );
 
@@ -41,40 +40,4 @@ public abstract class Behaviour implements RabbitNotifier
     public abstract void clearMemberVariables();
 
 
-
-    @Override
-    public void registerObserver( RabbitObserver observer, BehaviourExecutor behaviourExecutor)
-    {
-        ArrayList<RabbitObserver> observers = behaviourExecutor.getObservers();
-        observers.add( observer );
-        behaviourExecutor.setObservers( observers );
-    }
-
-    @Override
-    public void removeObserver( RabbitObserver observer, BehaviourExecutor behaviourExecutor)
-    {
-        ArrayList<RabbitObserver> observers = behaviourExecutor.getObservers();
-        observers.remove( observer );
-        behaviourExecutor.setObservers( observers );
-    }
-
-    @Override
-    public void notifyDeath(BehaviourExecutor behaviourExecutor)
-    {
-        ArrayList<RabbitObserver> observers = behaviourExecutor.getObservers();
-        for ( RabbitObserver observer : observers )
-        {
-            observer.updateDeath( behaviourExecutor );
-        }
-    }
-
-    @Override
-    public void notifyExiting(BehaviourExecutor behaviourExecutor)
-    {
-        ArrayList<RabbitObserver> observers = behaviourExecutor.getObservers();
-        for ( RabbitObserver observer : observers )
-        {
-            observer.updateExiting( behaviourExecutor );
-        }
-    }
 }
